separate concerns of apis in new tab page
