+++ components/brave_new_tab_ui/BUILD.gn
@@ -15,11 +15,14 @@ transpile_web_ui("brave_new_tab_ui") {
     "components/privateTab/torContent.tsx",
     "components/privateTab/torTab.tsx",
     "components/app.tsx",
-    "constants/backgrounds.ts",
+    "data/backgrounds.ts",
     "constants/new_tab_types.ts",
     "reducers/index.ts",
     "reducers/new_tab_reducer.tsx",
-    "api.ts",
+    "api/topSites/grid.ts",
+    "api/dataFetch.ts",
+    "api/topSites/bookmarks.ts",
+    "api/topSites/dnd.ts",
     "brave_new_tab.html",
     "brave_new_tab.tsx",
     "storage.ts",
