+++ components/brave_new_tab_ui/api/topSites/grid.ts
@@ -2,52 +2,13 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this file,
  * You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-import { bindActionCreators } from 'redux'
+// API
+import * as dataFetchAPI from '../dataFetch'
+import * as bookmarksAPI from './bookmarks'
+import { getCharForSite } from '../../helpers/newTabUtils'
 
 // Utils
-import * as newTabActions from './actions/new_tab_actions'
-import { debounce } from '../common/debounce'
-import store from './store'
-
-let actions: any
-
-const getActions = () => {
-  if (actions) {
-    return actions
-  }
-  actions = bindActionCreators(newTabActions, store.dispatch.bind(store))
-  return actions
-}
-
-/**
- * Obtains a letter / char that represents the current site
- */
-const getLetterFromSite = (site: NewTab.Site) => {
-  let name
-  try {
-    name = new window.URL(site.url || '').hostname
-  } catch (e) {
-    console.warn('getLetterFromSite', { url: site.url || '' })
-  }
-  name = site.title || name || '?'
-  return name.charAt(0).toUpperCase()
-}
-
-/**
- * Obtains the top sites and submits an action with the results
- */
-export const fetchTopSites = () => {
-  chrome.topSites.get((topSites: NewTab.Site[]) => getActions().topSitesDataUpdated(topSites || []))
-}
-
-/**
- * Obtains the URL's bookmark info and calls an action with the result
- */
-export const fetchBookmarkInfo = (url: string) => {
-  chrome.bookmarks.search(url.replace(/^https?:\/\//, ''),
-    (bookmarkTreeNodes) => getActions().bookmarkInfoAvailable(url, bookmarkTreeNodes[0])
-  )
-}
+import { debounce } from '../../../common/debounce'
 
 export const getGridSites = (state: NewTab.State, checkBookmarkInfo?: boolean) => {
   const sizeToCount = { large: 18, medium: 12, small: 6 }
@@ -73,13 +34,13 @@ export const getGridSites = (state: NewTab.State, checkBookmarkInfo?: boolean) =
 
   gridSites = gridSites.slice(0, count)
   gridSites.forEach((gridSite: NewTab.Site) => {
-    gridSite.letter = getLetterFromSite(gridSite)
+    gridSite.letter = getCharForSite(gridSite)
     gridSite.thumb = `chrome://thumb/${gridSite.url}`
     gridSite.favicon = `chrome://favicon/size/64@1x/${gridSite.url}`
     gridSite.bookmarked = state.bookmarks[gridSite.url]
 
     if (checkBookmarkInfo && !gridSite.bookmarked) {
-      fetchBookmarkInfo(gridSite.url)
+      bookmarksAPI.fetchBookmarkInfo(gridSite.url)
     }
   })
   return gridSites
@@ -89,5 +50,5 @@ export const getGridSites = (state: NewTab.State, checkBookmarkInfo?: boolean) =
  * Calculates the top sites grid and calls an action with the results
  */
 export const calculateGridSites = debounce((state: NewTab.State) => {
-  getActions().gridSitesUpdated(getGridSites(state, true))
+  dataFetchAPI.getActions().gridSitesUpdated(getGridSites(state, true))
 }, 10)
