+++ components/brave_new_tab_ui/data/backgrounds.ts
