+++ components/brave_new_tab_ui/reducers/new_tab_reducer.tsx
@@ -7,75 +7,19 @@ import { Reducer } from 'redux'
 // Constants
 import { types } from '../constants/new_tab_types'
 
-// Utils
+// API
+import * as gridAPI from '../api/topSites/grid'
+import * as dataFetchAPI from '../api/dataFetch'
+import * as bookmarksAPI from '../api/topSites/bookmarks'
+import * as dndAPI from '../api/topSites/dnd'
 import * as storage from '../storage'
-import { fetchBookmarkInfo, getGridSites, calculateGridSites, fetchTopSites } from '../api'
-
-const updateBookmarkInfo = (state: NewTab.State, url: string, bookmarkTreeNode?: NewTab.Bookmark) => {
-  const bookmarks = state.bookmarks
-  const gridSites = state.gridSites.slice()
-  const topSites = state.topSites.slice()
-  const pinnedTopSites = state.pinnedTopSites.slice()
-  // The default empty object is just to avoid null checks below
-  const gridSite: Partial<NewTab.Site> = gridSites.find((s) => s.url === url) || {}
-  const topSite: Partial<NewTab.Site> = topSites.find((s) => s.url === url) || {}
-  const pinnedTopSite: Partial<NewTab.Site> = pinnedTopSites.find((s) => s.url === url) || {}
-
-  if (bookmarkTreeNode) {
-    bookmarks[url] = bookmarkTreeNode
-    gridSite.bookmarked = topSite.bookmarked = pinnedTopSite.bookmarked = bookmarkTreeNode
-  } else {
-    delete bookmarks[url]
-    gridSite.bookmarked = topSite.bookmarked = pinnedTopSite.bookmarked = undefined
-  }
-  state = { ...state, bookmarks, gridSites }
-
-  return state
-}
-
-const onDraggedSite = (state: NewTab.State, url: string, destUrl: string) => {
-  const gridSitesWithoutPreview = getGridSites(state)
-  const currentPositionIndex = gridSitesWithoutPreview.findIndex(site => site.url === url)
-  const finalPositionIndex = gridSitesWithoutPreview.findIndex(site => site.url === destUrl)
-  let pinnedTopSites = state.pinnedTopSites.slice()
-
-  // A site that is not pinned yet will become pinned
-  const pinnedMovingSite = pinnedTopSites.find(site => site.url === url)
-  if (!pinnedMovingSite) {
-    const movingTopSite = Object.assign({}, gridSitesWithoutPreview.find(site => site.url === url))
-    movingTopSite.index = currentPositionIndex
-    movingTopSite.pinned = true
-    pinnedTopSites.push(movingTopSite)
-  }
-
-  pinnedTopSites = pinnedTopSites.map((pinnedTopSite) => {
-    pinnedTopSite = Object.assign({}, pinnedTopSite)
-    const currentIndex = pinnedTopSite.index
-    if (currentIndex === currentPositionIndex) {
-      pinnedTopSite.index = finalPositionIndex
-    } else if (currentIndex > currentPositionIndex && pinnedTopSite.index <= finalPositionIndex) {
-      pinnedTopSite.index = pinnedTopSite.index - 1
-    } else if (currentIndex < currentPositionIndex && pinnedTopSite.index >= finalPositionIndex) {
-      pinnedTopSite.index = pinnedTopSite.index + 1
-    }
-    return pinnedTopSite
-  })
-  state = { ...state, pinnedTopSites }
-  state = { ...state, gridSites: getGridSites(state) }
-  return state
-}
-
-const onDragEnd = (state: NewTab.State) => {
-  state = { ...state, gridSites: getGridSites(state) }
-  return state
-}
 
 export const newTabReducer: Reducer<NewTab.State | undefined> = (state: NewTab.State | undefined, action: any) => {
   if (state === undefined) {
     state = storage.load()
 
     setImmediate(() => {
-      fetchTopSites()
+      dataFetchAPI.fetchTopSites()
     })
   }
 
@@ -89,7 +33,7 @@ export const newTabReducer: Reducer<NewTab.State | undefined> = (state: NewTab.S
           title: topSite.title,
           url: topSite.url
         }, () => {
-          fetchBookmarkInfo(payload.url)
+          bookmarksAPI.fetchBookmarkInfo(payload.url)
         })
       }
       break
@@ -97,13 +41,13 @@ export const newTabReducer: Reducer<NewTab.State | undefined> = (state: NewTab.S
       const bookmarkInfo = state.bookmarks[payload.url]
       if (bookmarkInfo) {
         chrome.bookmarks.remove(bookmarkInfo.id, () => {
-          fetchBookmarkInfo(payload.url)
+          bookmarksAPI.fetchBookmarkInfo(payload.url)
         })
       }
       break
     case types.NEW_TAB_TOP_SITES_DATA_UPDATED:
       state = { ...state, topSites: payload.topSites }
-      calculateGridSites(state)
+      gridAPI.calculateGridSites(state)
       break
 
     case types.NEW_TAB_SITE_PINNED: {
@@ -118,7 +62,7 @@ export const newTabReducer: Reducer<NewTab.State | undefined> = (state: NewTab.S
         ...state,
         pinnedTopSites
       }
-      calculateGridSites(state)
+      gridAPI.calculateGridSites(state)
       break
     }
 
@@ -132,7 +76,7 @@ export const newTabReducer: Reducer<NewTab.State | undefined> = (state: NewTab.S
           pinnedTopSites
         }
       }
-      calculateGridSites(state)
+      gridAPI.calculateGridSites(state)
       break
 
     case types.NEW_TAB_SITE_IGNORED: {
@@ -144,7 +88,7 @@ export const newTabReducer: Reducer<NewTab.State | undefined> = (state: NewTab.S
         ignoredTopSites,
         showSiteRemovalNotification: true
       }
-      calculateGridSites(state)
+      gridAPI.calculateGridSites(state)
       break
     }
 
@@ -156,7 +100,7 @@ export const newTabReducer: Reducer<NewTab.State | undefined> = (state: NewTab.S
         ignoredTopSites,
         showSiteRemovalNotification: false
       }
-      calculateGridSites(state)
+      gridAPI.calculateGridSites(state)
       break
     }
 
@@ -166,7 +110,7 @@ export const newTabReducer: Reducer<NewTab.State | undefined> = (state: NewTab.S
         ignoredTopSites: [],
         showSiteRemovalNotification: false
       }
-      calculateGridSites(state)
+      gridAPI.calculateGridSites(state)
       break
 
     case types.NEW_TAB_HIDE_SITE_REMOVAL_NOTIFICATION:
@@ -177,15 +121,15 @@ export const newTabReducer: Reducer<NewTab.State | undefined> = (state: NewTab.S
       break
 
     case types.NEW_TAB_SITE_DRAGGED:
-      state = onDraggedSite(state, payload.fromUrl, payload.toUrl)
+      state = dndAPI.onDraggedSite(state, payload.fromUrl, payload.toUrl)
       break
 
     case types.NEW_TAB_SITE_DRAG_END:
-      state = onDragEnd(state)
+      state = dndAPI.onDragEnd(state)
       break
 
     case types.NEW_TAB_BOOKMARK_INFO_AVAILABLE:
-      state = updateBookmarkInfo(state, payload.queryUrl, payload.bookmarkTreeNode)
+      state = bookmarksAPI.updateBookmarkInfo(state, payload.queryUrl, payload.bookmarkTreeNode)
       break
 
     case types.NEW_TAB_GRID_SITES_UPDATED:
